+++ samples/sample-koa/src/app.ts
@@ -10,6 +10,7 @@ import {EnableDI} from "@nodeboot/di";
 import {EnableRepositories} from "@nodeboot/starter-persistence";
 import {EnableScheduling} from "@nodeboot/starter-scheduler";
 import {EnableComponentScan} from "@nodeboot/scan";
+import {EnableHttpClients} from "@nodeboot/starter-http";
 
 @EnableDI(Container)
 @EnableOpenApi()
@@ -18,6 +19,7 @@ import {EnableComponentScan} from "@nodeboot/scan";
 @EnableActuator()
 @EnableRepositories()
 @EnableScheduling()
+@EnableHttpClients()
 @EnableComponentScan()
 @NodeBootApplication()
 export class FactsServiceApp implements NodeBootApp {