+++ samples/sample-koa/src/services/users.service.ts
@@ -6,6 +6,7 @@ import {runOnTransactionCommit, runOnTransactionRollback, Transactional} from "@
 import {HttpError, NotFoundError} from "@nodeboot/error";
 import {ConfigService} from "@nodeboot/config";
 import {Users} from "../persistence/users.init";
+import {MicroserviceHttpClient} from "../clients/MicroserviceHttpClient";
 
 @Service()
 export class UserService {
@@ -13,8 +14,14 @@ export class UserService {
         private readonly logger: Logger,
         private readonly configService: ConfigService,
         private readonly userRepository: UserRepository,
+        private readonly httpClient: MicroserviceHttpClient,
     ) {
-        Users.forEach(user => this.userRepository.save(user));
+        // Workaround to load dummy users in the database
+        this.userRepository.count({}).then(existingUsers => {
+            if (!existingUsers) {
+                Users.forEach(user => this.userRepository.save(user));
+            }
+        });
     }
 
     public async findAllUser(): Promise<User[]> {
@@ -23,6 +30,13 @@ export class UserService {
         this.logger.info(`Reading node-boot.app.name from app-config.yam: ${appName}`);
 
         return this.userRepository.find();
+    }
+
+    public async findExternalUsers(): Promise<User[]> {
+        this.logger.info("Getting users from external service");
+        const result = await this.httpClient.get("/users");
+        this.logger.info(`Found ${result.data.length} users by calling external API`);
+        return result.data;
     }
 
     public async findWithCustomQuery(): Promise<User[]> {