+++ starters/http-client/src/adapter/index.ts
@@ -1,0 +1,1 @@
+export {HttpClientAdapter} from "./HttpClientAdapter";