+++ starters/http-client/src/client/types.ts
@@ -1,0 +1,7 @@
+import {AxiosRequestConfig} from "axios";
+
+export const HTTP_CLIENT_FEATURE = Symbol("HTTP-Client-Feature");
+
+export interface HttpClientConfig extends AxiosRequestConfig {
+    httpLogging?: boolean;
+}