+++ starters/http-client/src/decorator/index.ts
@@ -1,0 +1,2 @@
+export {EnableHttpClients} from "./EnableHttpClients";
+export {HttpClient} from "./HttpClient";