+++ samples/sample-express/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @nodeboot/express-sample
 
+## 2.0.0
+
+### Major Changes
+
+-   Add starter package for injectable HTTP clients
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/starter-http@2.0.0
+    -   @nodeboot/scan@2.0.0
+
 ## 1.1.1
 
 ### Patch Changes